requirements.txt
anthropic>=0.40.0
python-telegram-bot>=20.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
